+++ R/defaults.R
@@ -8,7 +8,9 @@ def_fonts <- if( Sys.info()["sysname"] == "Windows" ){
 
 flextable_global <- new.env(parent = emptyenv())
 default_flextable_settings <- list(
-  font.family = def_fonts, font.size = 11, font.color = "black",
+  font.family = def_fonts,
+  cs.family = def_fonts, eastasia.family = def_fonts, hansi.family = def_fonts,
+  font.size = 11, font.color = "black",
   text.align = "left", padding.bottom = 5, padding.top = 5,
   padding.left = 5, padding.right = 5,
   border.color = "#666666",
@@ -20,7 +22,7 @@ default_flextable_settings <- list(
   na_str  = "",
   fmt_date = "%Y-%m-%d", fmt_datetime = "%Y-%m-%d %H:%M:%S",
   fonts_ignore = FALSE,
-  extra_css = "caption {color: #777;margin-top: 10px;margin-bottom: 10px;text-align: center;}",
+  extra_css = "",
   theme_fun = "theme_booktabs",
   post_process_pdf = function(x) x,
   post_process_docx = function(x) x,
@@ -38,7 +40,17 @@ flextable_global$defaults <- default_flextable_settings
 #' are automatically applied to every flextable you produce.
 #' Use `set_flextable_defaults()` to override them. Use `init_flextable_defaults()`
 #' to re-init all values with the package defaults.
-#' @param font.family single character value specifying font name.
+#' @param font.family single character value. When format is Word, it specifies the font to
+#' be used to format characters in the Unicode range (U+0000-U+007F).
+#' @param cs.family optional and only for Word. Font to be used to format
+#' characters in a complex script Unicode range. For example, Arabic
+#' text might be displayed using the "Arial Unicode MS" font.
+#' @param eastasia.family optional and only for Word. Font to be used to
+#' format characters in an East Asian Unicode range. For example,
+#' Japanese text might be displayed using the "MS Mincho" font.
+#' @param hansi.family optional and only for Word. Font to be used to format
+#' characters in a Unicode range which does not fall into one of the
+#' other categories.
 #' @param font.size font size (in point) - 0 or positive integer value.
 #' @param font.color font color - a single character value specifying
 #' a valid color (e.g. "#000000" or "black").
@@ -46,6 +58,7 @@ flextable_global$defaults <- default_flextable_settings
 #' is one of 'left', 'right', 'center', 'justify'.
 #' @param padding.bottom,padding.top,padding.left,padding.right paragraph paddings - 0 or
 #' positive integer value.
+#' @param padding padding (shortcut for top, bottom, left and right padding)
 #' @param border.color border color - single character value
 #' (e.g. "#000000" or "black").
 #' @param background.color cell background color - a single character value specifying a
@@ -58,8 +71,7 @@ flextable_global$defaults <- default_flextable_settings
 #' @param fonts_ignore if TRUE, pdf-engine pdflatex can be used instead of
 #' xelatex or lualatex. If pdflatex is used, fonts will be ignored because they are
 #' not supported by pdflatex, whereas with the xelatex and lualatex engines they are.
-#' @param extra_css css instructions to be integrated with the table. By default, it
-#' contains only the style for captions.
+#' @param extra_css css instructions to be integrated with the table.
 #' @param theme_fun a single character value (the name of the theme function
 #' to be applied) or a theme function (input is a flextable, output is a flextable).
 #' @param post_process_pdf,post_process_docx,post_process_html,post_process_pptx Post-processing functions
@@ -85,11 +97,15 @@ flextable_global$defaults <- default_flextable_settings
 #'
 #' \if{html}{\figure{fig_set_flextable_defaults_2.png}{options: width=50\%}}
 set_flextable_defaults <- function(
-  font.family = NULL, font.size = NULL, font.color = NULL,
-  text.align = NULL, padding.bottom = NULL, padding.top = NULL,
+  font.family = NULL,
+  font.size = NULL, font.color = NULL,
+  text.align = NULL,
+  padding = NULL,
+  padding.bottom = NULL, padding.top = NULL,
   padding.left = NULL, padding.right = NULL,
   border.color = NULL, background.color = NULL,
   table.layout = NULL,
+  cs.family = NULL, eastasia.family = NULL, hansi.family = NULL,
   decimal.mark = NULL, big.mark = NULL, digits = NULL, na_str = NULL,
   fmt_date = NULL, fmt_datetime = NULL, extra_css = NULL,
   fonts_ignore = NULL, theme_fun = NULL,
@@ -101,8 +117,24 @@ set_flextable_defaults <- function(
 
   x <- list()
 
+  if( !is.null(padding) ){
+    if( is.null( padding.top) ) padding.top <- padding
+    if( is.null( padding.bottom) ) padding.bottom <- padding
+    if( is.null( padding.left) ) padding.left <- padding
+    if( is.null( padding.right) ) padding.right <- padding
+  }
+
   if( !is.null(font.family) ){
     x$font.family <- font.family
+  }
+  if( !is.null(cs.family) ){
+    x$cs.family <- cs.family
+  }
+  if( !is.null(eastasia.family) ){
+    x$eastasia.family <- eastasia.family
+  }
+  if( !is.null(hansi.family) ){
+    x$hansi.family <- hansi.family
   }
 
   if( !is.null(font.size) && is.numeric(font.size) && !(font.size<0) ){
@@ -223,7 +255,8 @@ get_flextable_defaults <- function(){
 print.flextable_defaults <- function(x, ...){
 
   message("## style properties\n")
-  styles <- c("font.family", "font.size", "font.color", "text.align", "padding.bottom",
+  styles <- c("font.family", "hansi.family", "eastasia.family", "cs.family",
+              "font.size", "font.color", "text.align", "padding.bottom",
     "padding.top", "padding.left", "padding.right", "border.color",
     "background.color")
   df <- data.frame(property = styles, value = unlist(x[styles]), stringsAsFactors = FALSE)
@@ -240,7 +273,7 @@ print.flextable_defaults <- function(x, ...){
   message("")
 
   message("## table.layout is:", x$table.layout, "\n")
-  message("## default theme is:", x$theme_fun, "\n")
+  if(is.character(x$theme_fun)) message("## default theme is:", x$theme_fun, "\n")
 
   message("## HTML specific:")
   message("extra_css:", x$extra_css)
@@ -287,6 +320,7 @@ fp_text_default <- function(color = flextable_global$defaults$font.color,
                             italic = FALSE,
                             underlined = FALSE,
                             font.family = flextable_global$defaults$font.family,
+                            cs.family = NULL, eastasia.family = NULL, hansi.family = NULL,
                             vertical.align = "baseline",
                             shading.color = "transparent"){
   fp_text(
@@ -296,6 +330,7 @@ fp_text_default <- function(color = flextable_global$defaults$font.color,
     italic = italic,
     underlined = underlined,
     font.family = font.family,
+    cs.family = cs.family, eastasia.family = eastasia.family, hansi.family = hansi.family,
     vertical.align = vertical.align,
     shading.color = shading.color
   )